+++ Example/ChidoriMenuExample/MenuDefinitions/MenuRegistry.swift
@@ -1,0 +1,21 @@
+//
+//  MenuRegistry.swift
+//  ChidoriMenuExample
+//
+//  Created by 秋星桥 on 1/19/25.
+//
+
+import ChidoriMenu
+import UIKit
+
+struct MenuDefinition {
+    let title: String
+    let menu: UIMenu
+}
+
+enum MenuRegistry {
+    static let allMenus: [MenuDefinition] = CoreFunctionalityTests.allTests +
+        MenuStructureTests.allTests +
+        AdvancedFeaturesTests.allTests +
+        ProductionExamples.allExamples
+}
